MAINT: Renamed csv_to_dict.pt to format_dict.py
This is to match the class name to the file name.
